Move more code and variables from rarch_state out to input_driver and menu_driver
